Separate into unique components
